pyproject.toml
[tool.poetry]
name = "exemplo-prisma"
version = "0.1.0"
description = ""
authors = ["gustavorteuber <gugareis135@gmail.com>"]
readme = "README.md"
packages = [{include = "exemplo_prisma"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
